+++ src/cljs/cxengage_javascript_sdk/next/modules.cljs
@@ -11,6 +11,7 @@
             [cxengage-javascript-sdk.next.authentication :as authentication]
             [cxengage-javascript-sdk.next.entities :as entities]
             [cxengage-javascript-sdk.next.session :as session]
+            [cxengage-javascript-sdk.next.contacts :as contacts]
             [cxengage-javascript-sdk.next.interaction :as interaction]
             [cxengage-javascript-sdk.next.sqs :as sqs]
             [cxengage-javascript-sdk.next.messaging :as messaging]
@@ -41,8 +42,9 @@
         session-module (session/map->SessionModule. (gen-new-initial-module-config comm<))
         interaction-module (interaction/map->InteractionModule. (gen-new-initial-module-config comm<))
         voice-module (voice/map->VoiceModule. (gen-new-initial-module-config comm<))
-        entities-module (entities/map->EntitiesModule. (gen-new-initial-module-config comm<))]
-    (doseq [module [auth-module session-module interaction-module voice-module entities-module]]
+        entities-module (entities/map->EntitiesModule. (gen-new-initial-module-config comm<))
+        contacts-module (contacts/map->ContactsModule. (gen-new-initial-module-config comm<))]
+    (doseq [module [auth-module session-module interaction-module voice-module entities-module contacts-module]]
       (start-internal-module module))))
 
 (defn start-session-dependant-modules [comm<]